YelpBackend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-pymongo>=3.0.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
]
